use ModalManager lib, refactor managers into their own thing.  not ideal to tie one-to-one with services as multiple services could leverage a manager
