+++ flask/app/services/__init__.py
@@ -1,8 +1,4 @@
-class ProxyManager(type):
-    def __getattr__(cls, key):
-        if cls.__manager__:
-            return getattr(cls.__manager__, key)
-        raise AttributeError(key)
+from model_manager import ProxyManager
 
 
 class Service(object):
