+++ flask/app/services/todo_service.py
@@ -1,12 +1,7 @@
-from app.lib.model_manager import SqlAlchemyModelManager
-from app.models.todo_models import Todo
+from app.managers import TodoManager
 from app.services import Service
 
 __all__ = ['TodoService']
-
-
-class TodoManager(SqlAlchemyModelManager):
-    __model__ = Todo
 
 
 class TodoService(Service):
