+++ flask/requirements.txt
@@ -2,3 +2,4 @@ ct-core-api==2.1.0b0
 Flask-Principal==0.4.0
 Flask-Admin==1.4.2
 Flask-Mail==0.9.1
+ModelManager==0.1.1
